# api-docs.json
{
	"swagger": "2.0",
	"info": {
		"description": "Api Documentation",
		"version": "1.0",
		"title": "Api Documentation",
		"termsOfService": "urn:tos",
		"contact": {
			
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "localhost:9090",
	"basePath": "/",
	"tags": [
		{
			"name": "ticket-controller",
			"description": "Ticket Controller"
		}
	],
	"paths": {
		"/bookticket": {
			"post": {
				"tags": [
					"ticket-controller"
				],
				"summary": "bookTicket",
				"operationId": "bookTicketUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "passenger",
						"description": "passenger",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Passenger"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Ticket"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/getTicket/{ticketNum}": {
			"get": {
				"tags": [
					"ticket-controller"
				],
				"summary": "getTicket",
				"operationId": "getTicketUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "ticketNum",
						"in": "path",
						"description": "ticketNum",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Ticket"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		}
	},
	"definitions": {
		"Passenger": {
			"type": "object",
			"properties": {
				"email": {
					"type": "string"
				},
				"from": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"to": {
					"type": "string"
				},
				"trainNo": {
					"type": "integer",
					"format": "int64"
				}
			},
			"title": "Passenger"
		},
		"Ticket": {
			"type": "object",
			"properties": {
				"email": {
					"type": "string"
				},
				"from": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"price": {
					"type": "number",
					"format": "double"
				},
				"ticketNum": {
					"type": "string"
				},
				"to": {
					"type": "string"
				},
				"trainNo": {
					"type": "integer",
					"format": "int64"
				}
			},
			"title": "Ticket"
		}
	}
}